Fixed: Use the correct type size with Svg_FromDef()

+++ doomsday/engine/portable/include/svg.h
@@ -66,6 +66,6 @@ svgid_t Svg_UniqueId(Svg* svg);
  *
  * @return  Newly created Svg instance if definition was valid else @a NULL
  */
-Svg* Svg_FromDef(svgid_t uniqueId, const def_svgline_t* lines, size_t numLines);
+Svg* Svg_FromDef(svgid_t uniqueId, const def_svgline_t* lines, uint numLines);
 
 #endif /* LIBDENG_SVG_H */
